# Upgrade version of typescript-eslint/parser to 2.15.0