--- a/eslint-bridge/package.json
+++ b/eslint-bridge/package.json
@@ -30,7 +30,7 @@
     "@types/estree": "0.0.39",
     "@types/express": "4.16.0",
     "@types/jest": "23.3.2",
-    "@types/node": "^13.1.4",
+    "@types/node": "13.1.4",
     "fs-extra": "7.0.0",
     "jest": "24.9.0",
     "jest-sonar-reporter": "1.3.0",