--- a/its/ruling/src/test/expected/ts/TypeScript/typescript-S3696.json
+++ b/its/ruling/src/test/expected/ts/TypeScript/typescript-S3696.json
@@ -2,4 +2,11 @@
 'TypeScript:scripts/importDefinitelyTypedTests/importDefinitelyTypedTests.ts':[
 21,
 ],
+'TypeScript:src/harness/harnessLanguageService.ts':[
+372,
+],
+'TypeScript:src/services/services.ts':[
+1008,
+1038,
+],
 }