--- a/its/ruling/src/test/expected/ts/desktop/typescript-S4325.json
+++ b/its/ruling/src/test/expected/ts/desktop/typescript-S4325.json
@@ -18,7 +18,6 @@
 100,
 ],
 'desktop:app/src/shared-process/repositories-store.ts':[
-34,
 184,
 ],
 'desktop:app/src/ui/diff/code-mirror-host.tsx':[