--- a/its/ruling/src/test/expected/ts/postgraphql/typescript-S3696.json
+++ b/its/ruling/src/test/expected/ts/postgraphql/typescript-S3696.json
@@ -1,0 +1,5 @@
+{
+'postgraphql:src/__tests__/utils/withPgClient.ts':[
+22,
+],
+}