Fixes #1846 

10 new issues from typescript:S3696 (Expected an error object to be thrown.), seem to be TP
1 missing issue typescript:S4325 (Redundant casts and not-null assertions should be avoided), fixed FP

The performance was not improved dramatically though

Timing of sensor time for typescript ruling 

on master

```
INFO: Sensor TypeScript analysis [javascript] (done) | time=30506ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=27790ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=101573ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=69421ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=13802ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=6326ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=2825ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=14432ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=18600ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=2954ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=39660ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=49207ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=92588ms
```

on this branch 

```
INFO: Sensor TypeScript analysis [javascript] (done) | time=30527ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=26201ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=83796ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=63314ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=14945ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=7234ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=3478ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=15635ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=18611ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=3535ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=44748ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=42505ms
INFO: Sensor TypeScript analysis [javascript] (done) | time=97704ms
```